BGDIINF_SB-1792: changed docker registry to AWS ECR
Also use a docker tag with user name and git hash short in order to not
overwrite images on the repo.

NOTE: for the docker login to ecr we need an aws cli version 2 or
superior, by default in code only older version of aws cli is available
therefore we install a new one.