BGDIINF_SB-1792: Fixed makefile dependencies
The ci-check-format due to a dependency did a `dev` target which could
use then other packages version as in the pipfile.lock.

To fix this dependency has been simplified with a draw back that know
the user has to manually first run `make dev` to create and update the
dependencies prior to do any other make target.