BGDIINF_SB-1304: Reduced the worker timeout
Somehow sometimes the GetCapabilities request hangs during the query,
the request is then killed by the gunicorn worker and automatically
restarted. I could not find the reason nor a workaround this issue,
which is difficult to understand because of the lac of logs. But
reducing the worker timeout avoid having the request taking more than 60
seconds to be answered.