BGDIINF_SB-1304: Fixed GetCapabilities 308 Moved Permanently answer
All the GetCapabilities route with defaults where returning a 308 Moved
Permanently to the simplest form when called with the default values.
This is how flask works with the default see
https://flask.palletsprojects.com/en/2.0.x/api/#url-route-registrations

Therefore remove the defaults in add_url_rule and set the optional epsg
and lang in url as None. Then defaults are handled like for the
/1.0.0/WMTSCapabilities.xml endpoint.