BGDIINF_SB-2038: Allow slow GetTile request
Some Tile generation are very slow (more than 30 seconds), therefore a
wsgi timeout of 10 seconds is too short for those tiles. However this 10
seconds timeout was a workaround for another unknown issue with the
GetCapabilities where the DB query sometimes hang indefinitely until the
wsgi timeout was reached. The workaround was to reduce this timeout in
order to automatically retry in this case without waiting too much time.
We need here to find another workaround for this GetCapabilities issue.