BGDIINF_SB-2118: Implement S3 2nd level cache check
Now we check first if a tile is present on the 2nd level cache on S3. By
default putting the tile on S3 cache is done synchronously after the
request has been set back to the user. But this behaviour can still be
change by envrionment variable to be fully syncrhone or fully
asynchrone.