BGDIINF_SB-2118: Removed Celery
Now S3 caching is fixed and is done using the Flask call_on_close CB.
The whole async with Celery/Rabbitmq has been removed. Due to this
unittest had to be quite reworked because it was not easy to test the
call_on_close CB which is not called using Flask test client.