BGDIINF_SB-2118: Added custom header for S3 write events
This custom header is set each time a S3 write is done with the request.
This allow to better analyze performance and cache behavior with K6.
With k6 we can then draw graph about request that triggered S3 write and
also compare performance in comparison to request that don't trigger a
S3 write.