BGDIINF_SB-2765: Improved logging in case of errors - #patch
Removed the exc_info=True in the S3 helper logs as this error happens quite a lot
and is correclty handled. This flag should be kept for non handled errors.

Also add the response.status_code field in error message so that we can
filter for those error message by status code. Note on PROD we only log level
from WARNING and above.