correcting the max zoom in the tilegridset webmercator for respect of the latitude of Swwitzerland.
Most layer in the BGDI/NGDI are only available in the tiny area of Switzerland
and defined in this area (resolution, zoom). When using global TileMatrix set like
the Spherical WebMercator one, where these paramters are defined at the Equator,
we have to correct them to match those of the area of interest.